backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.17.0

# Network monitoring
scapy>=2.5.0
netaddr>=0.8.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
requests>=2.28.0
websockets>=11.0.0

# Optional dependencies (comment out if causing issues)
# scikit-learn>=1.0.0
# tensorflow>=2.12.0
# redis>=4.5.0
# pymongo>=4.3.0
# aiohttp>=3.8.0
# beautifulsoup4>=4.11.0
# geopy>=2.3.0
# psycopg2-binary>=2.9.5
# asyncpg>=0.27.0
